refactor(cli): move Android Java/Gradle checks to a dedicated module
Extract the Java detection and Java/Gradle compatibility checks out of
android/mod.rs into android/java.rs, and recommend the exact supported
Java ceiling in the warning instead of a stale LTS list. Addresses @Legend-Master's feedback.
